Move some files to InstructionOrExpression folder in newIDE
