+++ newIDE/app/src/EventsSheet/InstructionEditor/InstructionOrExpressionSelector/ExpressionSelector.js
@@ -1,12 +1,15 @@
 // @flow
 import React, { Component } from 'react';
-import { enumerateExpressions } from './EnumerateExpressions';
+import { enumerateExpressions } from '../../../InstructionOrExpression/EnumerateExpressions';
 import InstructionOrExpressionSelector from './index';
-import { createTree, type InstructionOrExpressionTreeNode } from './CreateTree';
+import {
+  createTree,
+  type InstructionOrExpressionTreeNode,
+} from '../../../InstructionOrExpression/CreateTree';
 import {
   type EnumeratedInstructionOrExpressionMetadata,
   filterEnumeratedInstructionOrExpressionMetadataByScope,
-} from './EnumeratedInstructionOrExpressionMetadata.js';
+} from '../../../InstructionOrExpression/EnumeratedInstructionOrExpressionMetadata.js';
 import { type EventsScope } from '../../EventsScope.flow';
 
 type Props = {|
