+++ newIDE/app/src/EventsSheet/InstructionEditor/SelectorListItems/Keys.js
@@ -3,7 +3,7 @@ import type {
   GroupWithContext,
   ObjectWithContext,
 } from '../../../ObjectsList/EnumerateObjects';
-import { type EnumeratedInstructionOrExpressionMetadata } from '../InstructionOrExpressionSelector/EnumeratedInstructionOrExpressionMetadata.js';
+import { type EnumeratedInstructionOrExpressionMetadata } from '../../../InstructionOrExpression/EnumeratedInstructionOrExpressionMetadata.js';
 
 // ListItem created must have consistent keys that include their type
 // (object, group, instruction) to allow them to be living
