+++ newIDE/app/src/EventsSheet/ParameterFields/GenericExpressionField/FormatExpressionCall.js
@@ -1,5 +1,5 @@
 //@flow
-import { type EnumeratedInstructionOrExpressionMetadata } from '../../InstructionEditor/InstructionOrExpressionSelector/EnumeratedInstructionOrExpressionMetadata.js';
+import { type EnumeratedInstructionOrExpressionMetadata } from '../../../InstructionOrExpression/EnumeratedInstructionOrExpressionMetadata.js';
 import { type ParameterValues } from './ExpressionParametersEditorDialog';
 
 const filterOutCodeOnlyParameters = (
