+++ newIDE/app/src/InstructionOrExpression/EnumerateInstructions.js
