+++ newIDE/app/src/InstructionOrExpression/EnumerateInstructions.spec.js
